backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
attrs==25.4.0
beautifulsoup4==4.14.3
cattrs==25.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
curl_cffi==0.13.0
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.0
feedparser==6.0.12
frozendict==2.4.7
groq==1.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
multitasking==0.0.12
numpy==2.0.2
pandas==2.3.3
peewee==3.19.0
platformdirs==4.4.0
protobuf==6.33.4
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-cache==1.2.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
url-normalize==2.2.1
urllib3==2.6.3
uvicorn==0.39.0
webencodings==0.5.1
websockets==15.0.1
yfinance==0.2.36
